backend/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
billiard==3.6.4.0
blinker==1.5
Brotli==1.0.9
cachelib==0.9.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.6.0
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
fonttools==4.39.3
greenlet==2.0.1
html5lib==1.1
idna==3.4
imageio==2.24.0
importlib-metadata==5.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.1
numpy==1.24.1
pandas==2.0.0
passlib==1.7.4
pdfkit==1.0.0
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
premailer==3.10.0
prompt-toolkit==3.0.38
pycparser==2.21
pydyf==0.6.0
pyphen==0.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7
redis==4.5.4
reportlab==3.6.12
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.45
tinycss2==1.2.1
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
weasyprint==58.1
webencodings==0.5.1
Werkzeug==2.2.2
WTForms==3.0.1
yagmail==0.15.293
zipp==3.11.0
zopfli==0.2.2
